Move logic about metrics into separated code file
